:recycle: 컴포넌트별 역할 명확하게 수정
